[HUDI-3161] Add Call Produce Command for spark sql
