+++ hudi-spark-datasource/hudi-spark/src/main/antlr4/org/apache/hudi/spark/sql/parser/HoodieSqlCommon.g4
@@ -14,7 +14,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-grammar HoodieSqlCommon;
+
+ grammar HoodieSqlCommon;
+
+ @members {
+    /**
+     * When false, INTERSECT is given the greater precedence over the other set
+     * operations (UNION, EXCEPT and MINUS) as per the SQL standard.
+     */
+    public boolean legacy_setops_precedence_enbled = false;
+
+    /**
+     * Verify whether current token is a valid decimal token (which contains dot).
+     * Returns true if the character that follows the token is not a digit or letter or underscore.
+     *
+     * For example:
+     * For char stream "2.3", "2." is not a valid decimal token, because it is followed by digit '3'.
+     * For char stream "2.3_", "2.3" is not a valid decimal token, because it is followed by '_'.
+     * For char stream "2.3W", "2.3" is not a valid decimal token, because it is followed by 'W'.
+     * For char stream "12.0D 34.E2+0.12 "  12.0D is a valid decimal token because it is followed
+     * by a space. 34.E2 is a valid decimal token because it is followed by symbol '+'
+     * which is not a digit or letter or underscore.
+     */
+    public boolean isValidDecimal() {
+      int nextChar = _input.LA(1);
+      if (nextChar >= 'A' && nextChar <= 'Z' || nextChar >= '0' && nextChar <= '9' ||
+        nextChar == '_') {
+        return false;
+      } else {
+        return true;
+      }
+    }
+  }
 
  singleStatement
     : statement EOF
@@ -22,32 +53,115 @@ grammar HoodieSqlCommon;
 
  statement
     : compactionStatement                                              #compactionCommand
+    | callStatement                                                    #callProcedure
     | .*?                                                              #passThrough
     ;
 
  compactionStatement
-    : operation = (RUN | SCHEDULE) COMPACTION  ON tableIdentifier (AT instantTimestamp = NUMBER)?    #compactionOnTable
-    | operation = (RUN | SCHEDULE) COMPACTION  ON path = STRING   (AT instantTimestamp = NUMBER)?    #compactionOnPath
-    | SHOW COMPACTION  ON tableIdentifier (LIMIT limit = NUMBER)?                             #showCompactionOnTable
-    | SHOW COMPACTION  ON path = STRING (LIMIT limit = NUMBER)?                               #showCompactionOnPath
+    : operation = (RUN | SCHEDULE) COMPACTION  ON tableIdentifier (AT instantTimestamp = INTEGER_VALUE)?    #compactionOnTable
+    | operation = (RUN | SCHEDULE) COMPACTION  ON path = STRING   (AT instantTimestamp = INTEGER_VALUE)?    #compactionOnPath
+    | SHOW COMPACTION  ON tableIdentifier (LIMIT limit = INTEGER_VALUE)?                             #showCompactionOnTable
+    | SHOW COMPACTION  ON path = STRING (LIMIT limit = INTEGER_VALUE)?                               #showCompactionOnPath
     ;
 
  tableIdentifier
     : (db=IDENTIFIER '.')? table=IDENTIFIER
     ;
 
+ callStatement
+   : CALL multipartIdentifier '(' (callArgument (',' callArgument)*)? ')'  #call
+   ;
+
+ multipartIdentifier
+     : parts+=identifier ('.' parts+=identifier)*
+     ;
+
+ callExpression
+     : constant
+     | stringMap
+     ;
+
+ callArgument
+     : callExpression                        #positionalArgument
+     | identifier '=>' callExpression        #namedArgument
+     ;
+
+ stringMap
+     : MAP '(' constant (',' constant)* ')'
+     ;
+
+ identifier
+     : strictIdentifier
+     ;
+
+ strictIdentifier
+     : IDENTIFIER             #unquotedIdentifier
+     | quotedIdentifier       #quotedIdentifierAlternative
+     | nonReserved            #unquotedIdentifier
+     ;
+
+ quotedIdentifier
+     : BACKQUOTED_IDENTIFIER
+     ;
+
+ constant
+     : NULL                                                                                     #nullLiteral
+     | interval                                                                                 #intervalLiteral
+     | identifier STRING                                                                        #typeConstructor
+     | number                                                                                   #numericLiteral
+     | booleanValue                                                                             #booleanLiteral
+     | STRING+                                                                                  #stringLiteral
+     ;
+
+ booleanValue
+      : TRUE | FALSE
+      ;
+
+ interval
+     : INTERVAL intervalField*
+     ;
+
+ intervalField
+     : value=intervalValue unit=identifier (TO to=identifier)?
+     ;
+
+ intervalValue
+     : (PLUS | MINUS)? (INTEGER_VALUE | DECIMAL_VALUE)
+     | STRING
+     ;
+
+ number
+     : MINUS? INTEGER_VALUE            #integerLiteral
+     | MINUS? DECIMAL_VALUE            #decimalLiteral
+     | MINUS? BIGINT_LITERAL           #bigIntLiteral
+     | MINUS? SMALLINT_LITERAL         #smallIntLiteral
+     | MINUS? TINYINT_LITERAL          #tinyIntLiteral
+     | MINUS? DOUBLE_LITERAL           #doubleLiteral
+     | MINUS? BIGDECIMAL_LITERAL       #bigDecimalLiteral
+     ;
+
+ nonReserved
+     : CALL | COMPACTION | RUN | SCHEDULE | ON | SHOW | LIMIT
+     ;
+
  ALL: 'ALL';
  AT: 'AT';
+ CALL: 'CALL';
  COMPACTION: 'COMPACTION';
  RUN: 'RUN';
  SCHEDULE: 'SCHEDULE';
  ON: 'ON';
  SHOW: 'SHOW';
  LIMIT: 'LIMIT';
+ MAP: 'MAP';
+ NULL: 'NULL';
+ TRUE: 'TRUE';
+ FALSE: 'FALSE';
+ INTERVAL: 'INTERVAL';
+ TO: 'TO';
 
- NUMBER
-    : DIGIT+
-    ;
+ PLUS: '+';
+ MINUS: '-';
 
  IDENTIFIER
     : (LETTER | DIGIT | '_')+
@@ -58,7 +172,49 @@ grammar HoodieSqlCommon;
     | '"' ( ~('"'|'\\') | ('\\' .) )* '"'
     ;
 
+ BACKQUOTED_IDENTIFIER
+     : '`' ( ~'`' | '``' )* '`'
+     ;
 
+ BIGINT_LITERAL
+     : DIGIT+ 'L'
+     ;
+
+ SMALLINT_LITERAL
+     : DIGIT+ 'S'
+     ;
+
+ TINYINT_LITERAL
+     : DIGIT+ 'Y'
+     ;
+
+ INTEGER_VALUE
+     : DIGIT+
+     ;
+
+ DECIMAL_VALUE
+     : DIGIT+ EXPONENT
+     | DECIMAL_DIGITS EXPONENT? {isValidDecimal()}?
+     ;
+
+ DOUBLE_LITERAL
+     : DIGIT+ EXPONENT? 'D'
+     | DECIMAL_DIGITS EXPONENT? 'D' {isValidDecimal()}?
+     ;
+
+ BIGDECIMAL_LITERAL
+     : DIGIT+ EXPONENT? 'BD'
+     | DECIMAL_DIGITS EXPONENT? 'BD' {isValidDecimal()}?
+     ;
+
+ fragment DECIMAL_DIGITS
+     : DIGIT+ '.' DIGIT*
+     | '.' DIGIT+
+     ;
+
+ fragment EXPONENT
+     : 'E' [+-]? DIGIT+
+     ;
 
  fragment DIGIT
      : [0-9]
